fixes for compiler warnings on solaris

+++ pam_ssh_agent_auth/authfd.c
@@ -116,7 +116,7 @@ ssh_get_authentication_socket(uid_t uid)
     /* Advisory only; seteuid ensures no race condition; but will only log if we see EACCES */
     if( stat(authsocket,&sock_st) == 0) {
         if(uid != 0 && sock_st.st_uid != uid) {
-            fatal("uid %d attempted to open an agent socket owned by uid %d", (int) uid, (int) sock_st.st_uid);
+            fatal("uid %lu attempted to open an agent socket owned by uid %lu", (unsigned long) uid, (unsigned long) sock_st.st_uid);
             return -1;
         }
     }
@@ -148,7 +148,7 @@ ssh_get_authentication_socket(uid_t uid)
 	if (connect(sock, (struct sockaddr *)&sunaddr, sizeof sunaddr) < 0) {
 		close(sock);
         if(errno == EACCES)
-            fatal("MAJOR SECURITY WARNING: uid %d made a deliberate and malicious attempt to open an agent socket owned by another user", (int) uid);
+            fatal("MAJOR SECURITY WARNING: uid %lu made a deliberate and malicious attempt to open an agent socket owned by another user", (unsigned long) uid);
 		return -1;
 	}
